Add map_groups
