+++ py-polars/polars/__init__.py
@@ -121,6 +121,7 @@ from polars.functions import (
     last,
     lit,
     map,
+    map_groups,
     max,
     max_horizontal,
     mean,
@@ -333,6 +334,7 @@ __all__ = [
     "last",
     "lit",
     "map",
+    "map_groups",
     "mean",
     "median",
     "n_unique",
