+++ py-polars/polars/functions/__init__.py
@@ -50,6 +50,7 @@ from polars.functions.lazy import (
     implode,
     last,
     map,
+    map_groups,
     mean,
     median,
     n_unique,
@@ -138,6 +139,7 @@ __all__ = [
     "last",
     "lit",
     "map",
+    "map_groups",
     "mean",
     "median",
     "n_unique",
