+++ py-polars/polars/lazyframe/group_by.py
@@ -5,6 +5,7 @@ from typing import TYPE_CHECKING, Callable, Iterable
 from polars import functions as F
 from polars.utils._parse_expr_input import parse_as_list_of_expressions
 from polars.utils._wrap import wrap_ldf
+from polars.utils.deprecation import deprecate_renamed_function
 
 if TYPE_CHECKING:
     from polars import DataFrame, LazyFrame
@@ -144,7 +145,7 @@ class LazyGroupBy:
         pyexprs = parse_as_list_of_expressions(*aggs, **named_aggs)
         return wrap_ldf(self.lgb.agg(pyexprs))
 
-    def apply(
+    def map_groups(
         self,
         function: Callable[[DataFrame], DataFrame],
         schema: SchemaDict | None,
@@ -175,7 +176,6 @@ class LazyGroupBy:
             Schema of the output function. This has to be known statically. If the
             given schema is incorrect, this is a bug in the caller's query and may
             lead to errors. If set to None, polars assumes the schema is unchanged.
-
 
         Examples
         --------
@@ -229,7 +229,7 @@ class LazyGroupBy:
         ... )  # doctest: +IGNORE_RESULT
 
         """
-        return wrap_ldf(self.lgb.apply(function, schema))
+        return wrap_ldf(self.lgb.map_groups(function, schema))
 
     def head(self, n: int = 5) -> LazyFrame:
         """
@@ -649,3 +649,27 @@ class LazyGroupBy:
 
         """
         return self.agg(F.all().sum())
+
+    @deprecate_renamed_function("map_groups", version="0.19.0")
+    def apply(
+        self,
+        function: Callable[[DataFrame], DataFrame],
+        schema: SchemaDict | None,
+    ) -> LazyFrame:
+        """
+        Apply a custom/user-defined function (UDF) over the groups as a new DataFrame.
+
+        .. deprecated:: 0.19.0
+            This method has been renamed to :func:`LazyGroupBy.map_groups`.
+
+        Parameters
+        ----------
+        function
+            Function to apply over each group of the `LazyFrame`.
+        schema
+            Schema of the output function. This has to be known statically. If the
+            given schema is incorrect, this is a bug in the caller's query and may
+            lead to errors. If set to None, polars assumes the schema is unchanged.
+
+        """
+        return self.map_groups(function, schema)
