+++ py-polars/tests/unit/operations/test_group_by_rolling.py
@@ -24,31 +24,6 @@ def good_agg_parameters() -> list[pl.Expr | list[pl.Expr]]:
     ]
 
 
-def test_group_by_rolling_map() -> None:
-    df = pl.DataFrame(
-        {
-            "a": [1, 2, 3, 4, 5],
-            "b": [1, 2, 3, 4, 5],
-        }
-    ).set_sorted("a")
-
-    def apply(df: pl.DataFrame) -> pl.DataFrame:
-        return df.select(
-            pl.col("a").min(),
-            pl.col("b").max(),
-        )
-
-    expected = pl.DataFrame(
-        [
-            pl.Series("a", [1, 1, 2, 3, 4], dtype=pl.Int64),
-            pl.Series("b", [1, 2, 3, 4, 5], dtype=pl.Int64),
-        ]
-    )
-
-    out = df.group_by_rolling("a", period="2i").apply(apply, schema=df.schema)
-    assert_frame_equal(out, expected)
-
-
 def test_rolling_group_by_overlapping_groups() -> None:
     # this first aggregates overlapping groups so they cannot be naively flattened
     df = pl.DataFrame({"a": [41, 60, 37, 51, 52, 39, 40]})
