+++ py-polars/tests/unit/test_empty.py
@@ -72,14 +72,6 @@ def test_empty_9137() -> None:
     assert out.dtypes == [pl.Float32, pl.Float32]
 
 
-def test_empty_group_by_apply_err() -> None:
-    df = pl.DataFrame(schema={"x": pl.Int64})
-    with pytest.raises(
-        pl.ComputeError, match=r"cannot group_by \+ apply on empty 'DataFrame'"
-    ):
-        df.group_by("x").apply(lambda x: x)
-
-
 def test_empty_list_namespace_output_9585() -> None:
     dtype = pl.List(pl.Utf8)
     names = ["sort", "unique", "head", "tail", "shift", "reverse"]
